app/dna_utils.py
import zlib
from Crypto.Cipher import AES
from Crypto.Util.Padding import pad, unpad

def compress_data(data):
    return zlib.compress(data)

def encrypt_data(data, key):
    cipher = AES.new(key, AES.MODE_ECB)
    return cipher.encrypt(pad(data, AES.block_size))

def encode_to_dna(binary_data):
    dna_map = {'00': 'A', '01': 'C', '10': 'G', '11': 'T'}
    return ''.join(dna_map[binary_data[i:i+2]] for i in range(0, len(binary_data), 2))

def process_file(file):
    data = file.read()
    compressed_data = compress_data(data)
    key = b'Sixteen byte key'  # Replace with a secure key generation method
    encrypted_data = encrypt_data(compressed_data, key)
    binary = ''.join(format(byte, '08b') for byte in encrypted_data)
    return encode_to_dna(binary)

# Add functions for decoding, decrypting, and decompressing here
