requirements.txt
blinker==1.8.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
MarkupSafe==3.0.2
numpy==2.1.2
pillow==11.0.0
pycryptodome==3.21.0
scikit-learn==1.5.2
scipy==1.14.1
threadpoolctl==3.5.0
Werkzeug==3.0.4
